Detect typos in master password
It's really annoying to type the master password twice, every time we
want to create a new password or obtain an existing one. But it's quite
dangerous to not be able to detect typos, which could prevent recovery
of newly created passwords.

To protect against this, we ask for the master password twice the first
time it is entered. Once it has been entered successfully twice, we keep
a randomized hash of it in memory. The next time we ask the user to
enter the master password, we compare the result to the stored hash: if
they match, we don't need to enter the password again.